wip functional client handler
