+++ src/main/java/com/flightstats/hub/metrics/StatsDReporter.java
@@ -1,0 +1,8 @@
+package com.flightstats.hub.metrics;
+
+public class StatsDReporter {
+    private StatsDHandlers statsDHandlers;
+
+
+
+}
